refactor by CR

+++ table_engine/src/partition/rule/df_adapter/mod.rs
@@ -113,7 +113,7 @@ mod tests {
         ];
         let partition_key_refs = partition_keys.iter().collect::<Vec<_>>();
         let mut buf = BytesMut::new();
-        let expected = compute_partition(&partition_key_refs, partition_num as u64, &mut buf);
+        let expected = compute_partition(&partition_key_refs, partition_num, &mut buf);
 
         assert_eq!(partitions[0], expected);
 
@@ -233,8 +233,8 @@ mod tests {
         let partition_keys_2 = test_datums[1].clone();
         let partition_key_refs_2 = partition_keys_2.iter().collect::<Vec<_>>();
         let mut buf = BytesMut::new();
-        let expected_1 = compute_partition(&partition_key_refs_1, partition_num as u64, &mut buf);
-        let expected_2 = compute_partition(&partition_key_refs_2, partition_num as u64, &mut buf);
+        let expected_1 = compute_partition(&partition_key_refs_1, partition_num, &mut buf);
+        let expected_2 = compute_partition(&partition_key_refs_2, partition_num, &mut buf);
         let expecteds = vec![expected_1, expected_2];
 
         assert_eq!(partitions, expecteds);
